+++ example/lib/main.dart
@@ -4,4 +4,4 @@
 
 import 'package:example/src/view.dart';
 
-void main() => runApp(MyApp(key: const Key('MyApp')));
+void main() => runApp(const MyApp(key: Key('MyApp')));