+++ example/test/widget_test.dart
@@ -4,9 +4,11 @@
 
 /// The 'show' clause is not essential. Merely for your reference.
 
-import 'package:example/src/view.dart' show MyApp, UniqueKey;
+import 'package:example/src/view.dart';
 
-import '../integration_test/_test_example_app.dart' show integrationTesting;
+import '../integration_test/_test_example_app.dart';
+
+import 'test_listener.dart' show testsStateListener01;
 
 import '_unit_testing.dart' show unitTesting;
 
@@ -23,20 +25,30 @@ void testMyApp() {
 
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('test mvc_pattern', (WidgetTester tester) async {
-    // Tells the tester to build a UI based on the widget tree passed to it
-    await tester.pumpWidget(app);
+  testWidgets(
+    'test mvc_pattern',
+    (WidgetTester tester) async {
+      // Tells the tester to build a UI based on the widget tree passed to it
+      await tester.pumpWidget(app);
 
-    /// Flutter won’t automatically rebuild your widget in the test environment.
-    /// Use pump() or pumpAndSettle() to ask Flutter to rebuild the widget.
+      /// Flutter won’t automatically rebuild your widget in the test environment.
+      /// Use pump() or pumpAndSettle() to ask Flutter to rebuild the widget.
 
-    /// pumpAndSettle() waits for all animations to complete.
-    await tester.pumpAndSettle();
+      /// pumpAndSettle() waits for all animations to complete.
+      await tester.pumpAndSettle();
 
-    /// Preform integration first to set up
-    /// WidgetsBinding.instance is IntegrationTestWidgetsFlutterBinding
-    await integrationTesting(tester);
+      /// Tests Listener class
+      await testsStateListener01(tester);
 
-    await unitTesting(tester);
-  });
+      /// Preform integration first to set up
+      /// WidgetsBinding.instance is IntegrationTestWidgetsFlutterBinding
+      await integrationTesting(tester);
+
+      /// Testing the StateMVC, ControllerMVC, and ListenerMVC
+      await unitTesting(tester);
+
+      /// Reset the counter to zero on Page 1
+      await resetPage1Count(tester);
+    },
+  );
 }